// src/StreamMaster.UI/lib/api.d.ts
export interface components {
	schemas: {
// ...
		SDSettings: {
			/** Format: int32 */
			maxSubscribedLineups?: number;
			alternateSEFormat?: boolean;
			alternateLogoStyle?: string;
			/** Format: date-time */
			tokenErrorTimestamp?: string;
			artworkSize?: string;
			excludeCastAndCrew?: boolean;
			preferredLogoStyle?: string;
			prefixEpisodeTitle?: boolean;
			appendEpisodeDesc?: boolean;
			prefixEpisodeDescription?: boolean;
			episodeAppendProgramDescription?: boolean;
			sdEnabled?: boolean;
			/** Format: int32 */
			sdepgDays?: number;
			sdCountry?: string;
			sdPassword?: string;
			sdPostalCode?: string;
			sdStationIds?: components["schemas"]["StationIdLineup"][];
			headendsToView?: components["schemas"]["HeadendToView"][];
			sdUserName?: string;
			movieImages?: boolean;
			seasonImages?: boolean;
			seriesImages?: boolean;
			sportsImages?: boolean;
			episodeImages?: boolean;
			seriesPosterAspect?: string;
			moviePosterAspect?: string;
			xmltvAddFillerData?: boolean;
			/** Format: int32 */
			xmltvFillerProgramLength?: number;
			xmltvExtendedInfoInTitleDescriptions?: boolean;
			xmltvIncludeChannelNumbers?: boolean;
			xmltvSingleImage?: boolean;
			/** Format: date-time */
			sdTooManyRequestsSuspend?: string;
		};
// ...
	};
	responses: never;
	parameters: never;
	requestBodies: never;
	headers: never;
	pathItems: never;
}